fix bugs

+++ prepare.sh
@@ -19,7 +19,7 @@ for addr in ${TIKV}; do
     echo "  - host: ${addr}" >> ./data/topology.yaml
 done
 # connect to servers
-if ![ "${SSH_KEY}" ]; then
+if [ "${SSH_KEY}" = "" ]; then
     read -p "setup ssh? (y/n)" set_ssh
     if [ "${set_ssh}" = "y" ]; then
         ssh-keygen -t rsa -b 4096 -f ./data/key -P ""
@@ -31,31 +31,38 @@ fi
 # check connectivity
 for addr in ${SERVERS}; do
     addr=${USER_NAME}@${addr}
-    ssh ${addr} "sudo echo 'successful sudo'"
-    scp ./scripts/* ${addr}:~/
+    ssh ${SSH_FLAGS} ${addr} "sudo echo 'successful sudo'"
+    scp ${SSH_FLAGS} ./scripts/* ${addr}:~/
 done
 # check directory
 read -p "setup dir? (y/n)" set_dir
 if [ "${set_dir}" = "y" ]; then
     for addr in ${SERVERS}; do
         addr=${USER_NAME}@${addr}
-        echo "on ${add}:"
-        ssh ${addr} "lsblk"
-        read -p "which device?" device
-        read -p "mount name?" name
-        ssh ${addr} "mount.sh ${device} /${name} && sudo ln -s /${name} /data\
-            && sudo ln -s /${name} /backup && sudo chown -R ${USER_NAME}:${USER_NAME}\
-             /data && sudo chown -R ${USER_NAME}:${USER_NAME} /backup && sudo\
-             chown -R ${USER_NAME}:${USER_NAME} /${name}"
+        echo "on ${addr}:"
+        ssh ${SSH_FLAGS} ${addr} "lsblk"
+        read -p "mount dir? (y/n)" mount_dir
+        if [ "${mount_dir}" = "y" ]; then
+            read -p "which device?" device
+            read -p "mount name?" name
+            ssh ${SSH_FLAGS} ${addr} "~/mount.sh ${device} /${name} && sudo rm -rf /data \
+                && sudo rm -rf /backup && sudo ln -s /${name} /data \
+                && sudo ln -s /${name} /backup && sudo chown -R ${USER_NAME}:${USER_NAME} \
+                /data && sudo chown -R ${USER_NAME}:${USER_NAME} /backup && sudo \
+                chown -R ${USER_NAME}:${USER_NAME} /${name}"
+        else
+            ssh ${SSH_FLAGS} ${addr} "sudo mkdir -p /data && sudo mkdir -p /backup && sudo chown -R \
+                ${USER_NAME}:${USER_NAME} /data && sudo chown -R ${USER_NAME}:${USER_NAME} /backup"
+        fi
     done
 else
     for addr in ${SERVERS}; do
         addr=${USER_NAME}@${addr}
-        OWNER=$(ssh ${addr} "ls -ld ${DATA_DIR}" | awk '{print $3}')
+        OWNER=$(ssh ${SSH_FLAGS} ${addr} "ls -ld ${DATA_DIR}" | awk '{print $3}')
         if [ ${OWNER} != ${USER_NAME} ]; then
             echo "DATA_DIR owner isn't ${USER_NAME}" 
         fi
-        OWNER=$(ssh ${addr} "ls -ld ${BACKUP_DIR}" | awk '{print $3}')
+        OWNER=$(ssh ${SSH_FLAGS} ${addr} "ls -ld ${BACKUP_DIR}" | awk '{print $3}')
         if [ ${OWNER} != ${USER_NAME} ]; then
             echo "BACKUP_DIR owner isn't ${USER_NAME}" 
         fi
@@ -67,9 +74,9 @@ for addr in ${TIDB}; do
     echo "Host db${i}
         Hostname ${addr}
         User ${USER_NAME}
-        IdentityFile ${SSH_KEY}" >> ~/.ssh/config
-    if [ "${SSH_KEY}" ]; then
-	echo "IdentityFile ${SSH_KEY}" >> ~/.ssh/config
+        IdentityFile ${PRIVATE_SSH_KEY}" >> ~/.ssh/config
+    if [ "${PRIVATE_SSH_KEY}" ]; then
+	echo "IdentityFile ${PRIVATE_SSH_KEY}" >> ~/.ssh/config
     fi
     i=$((i+1))
 done
@@ -78,9 +85,9 @@ for addr in ${TIKV}; do
     echo "Host kv${i}
         Hostname ${addr}
         User ${USER_NAME}
-        IdentityFile ${SSH_KEY}" >> ~/.ssh/config
-    if [ "${SSH_KEY}" ]; then
-        echo "IdentityFile ${SSH_KEY}" >> ~/.ssh/config
+        IdentityFile ${PRIVATE_SSH_KEY}" >> ~/.ssh/config
+    if [ "${PRIVATE_SSH_KEY}" ]; then
+        echo "IdentityFile ${PRIVATE_SSH_KEY}" >> ~/.ssh/config
     fi
     i=$((i+1))
 done
